comment
I've got forms where I add a few fields at a time with a radio button and the first "submit" click doesn't seem to perform validation. Subsequent clicks do work, however. This *should* be handled by the mutation observer, but something is off. For now I'm working around this by manually reinitializing in the function that adds or removes inputs in a form.